requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
spotipy>=2.23.0
requests>=2.28.0
python-dotenv>=1.0.0
